requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.3.0
django-ratelimit==4.0.0
django-axes==6.0.0
django-allauth==0.57.0
django-csp==3.8
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.0
pyjwt==2.8.0
argon2-cffi==23.1.0
whitenoise==6.6.0
gunicorn==21.2.0
celery==5.3.4
django-celery-beat==2.5.0
drf-yasg==1.21.7
django-extensions==3.2.3
python-dotenv==1.0.0
django-filter==23.5
django-debug-toolbar==4.2.0